web: revamping the form in prep for the wizard
This commit updates three (!) of the 10 or so components that we
routinely use in our product.  These components are not currently
in use; I'm just building them, testing them with storybook and
providing continuous integration behavioral tests with Jest and
Chai.

The Storybook comes with complete interactivity and forms, and
there's a very neat transition for phones.